Parameter type clean-up.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk@3892 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ bogofilter/src/lexer.c
@@ -71,9 +71,9 @@ static void lexer_display_buffer(buff_t *buff)
 /* Check for lines wholly composed of printable characters as they can cause a scanner abort 
    "input buffer overflow, can't enlarge buffer because scanner uses REJECT"
 */
-static bool not_long_token(byte *buf, size_t count)
+static bool not_long_token(byte *buf, uint count)
 {
-    size_t i;
+    uint i;
     for (i=0; i < count; i += 1) {
 	unsigned char c = (unsigned char)buf[i];
 	if ((iscntrl(c) || isspace(c) || ispunct(c)) && (c != '_'))
